requirements.txt
Django==5.0.3
djangorestframework==3.14.0
django-cors-headers==4.3.1
opencv-python==4.8.1.78
face-recognition==1.3.0
PyMySQL==1.1.0
pyjwt==2.8.0
requests==2.31.0
openai==1.14.0
numpy==1.26.3